refactor: convert helpers to TS
