+++ src/helpers/processOperatorInput.ts
@@ -1,6 +1,7 @@
 import { appendOperator } from './appendOperator'
+import { AppState } from '../constants'
 
-export const processOperatorInput = (state, input) => {
+export const processOperatorInput = (state: AppState, input: string) => {
   return state.formula
     ? appendOperator(state, input)
     : { ...state, formula: '0' }
